Update pigLatin
Added code  for all the scenarios. Unit testing needs to be done more for checking casing of the letters. And the code just checks the words right now, also need to handle if string of words is entered. 
NOTE : vowelIndex class committed on the same branch must also be downloaded to have the code used.